Moved functions from script to core modules
